front end terminado

+++ proyectoIntegrador/src/main/java/com/backend/proyectoIntegrador/service/impl/OdontologoService.java
@@ -1,20 +1,19 @@
 package com.backend.proyectoIntegrador.service.impl;
 
 
-
 import com.backend.proyectoIntegrador.dto.OdontologoDto;
 import com.backend.proyectoIntegrador.entity.Odontologo;
-
+import com.backend.proyectoIntegrador.exceptions.ResourceNotFoundException;
 import com.backend.proyectoIntegrador.repository.OdontologoRepository;
 import com.backend.proyectoIntegrador.service.IOdontologoService;
+import com.backend.proyectoIntegrador.utils.JsonPrinter;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-
 
 
 @Service
@@ -34,59 +33,62 @@ public class OdontologoService implements IOdontologoService {
     public OdontologoDto buscarOdontologoPorId(Long id) {
         Odontologo odontologoBuscado = odontologoRepository.findById(id).orElse(null);
         OdontologoDto odontologoDto = null;
-        if (odontologoBuscado != null){
-            OdontologoDto odontologoDtoNuevo= objectMapper.convertValue(odontologoBuscado.getMatricula(), OdontologoDto.class);
-            odontologoBuscado.setMatricula(odontologoBuscado.getMatricula());
-            LOGGER.info("Odontologo encontrado: {}", odontologoDtoNuevo);
-        }else LOGGER.info("La matricula no se encuentra registrada");
+        if (odontologoBuscado != null) {
+            odontologoDto = objectMapper.convertValue(odontologoBuscado, OdontologoDto.class);
+            LOGGER.info("Dentist found: {}", JsonPrinter.toString(odontologoDto));
+        } else LOGGER.info("The dentist doesn´t exist");
 
         return odontologoDto;
     }
 
     @Override
     public List<OdontologoDto> listarOdontologos() {
-        List<Odontologo> odontologos = odontologoRepository.findAll();
-        List<OdontologoDto> odontologoDtos = odontologos.stream()
-                .map(odontologo ->{
-                    return new OdontologoDto(odontologo.getId(), odontologo.getMatricula(), odontologo.getNombre(), odontologo.getApellido());
+        return odontologoRepository.findAll()
+                .stream()
+                .map(odontologo -> {
+                    return new OdontologoDto(
+                            odontologo.getId(),
+                            odontologo.getMatricula(),
+                            odontologo.getNombre(),
+                            odontologo.getApellido());
                 })
                 .toList();
-        return odontologoDtos;
     }
 
     @Override
     public OdontologoDto registrarOdontologo(Odontologo odontologo) {
         Odontologo odontologoNuevo = odontologoRepository.save(odontologo);
-        OdontologoDto odontologoDtoNuevo = objectMapper.convertValue(odontologoNuevo.getMatricula(), OdontologoDto.class);
-        odontologoDtoNuevo.setMatricula(odontologo.getMatricula());
-
-        LOGGER.info("Nuevo odontologo registrado con exito: {}", odontologoDtoNuevo);
+        OdontologoDto odontologoDtoNuevo = objectMapper.convertValue(odontologoNuevo, OdontologoDto.class);
+        LOGGER.info("Dentist saved: {}", odontologoDtoNuevo);
 
         return odontologoDtoNuevo;
-
     }
 
     @Override
     public OdontologoDto actualizarOdontologo(Odontologo odontologo) {
-        Odontologo odontologoActualizar = odontologoRepository.findById(odontologo.getId()).orElse(null);
-        OdontologoDto odontologoActualizadoDto = null;
-        if (odontologoActualizar != null){
-            odontologoActualizar = odontologo;
-            odontologoRepository.save(odontologoActualizar);
-            odontologoActualizadoDto = objectMapper.convertValue(odontologoActualizar, OdontologoDto.class);
-            odontologoActualizadoDto.setMatricula(odontologo.getMatricula());
-            LOGGER.info("Paciente actualizado con exito: {}", odontologoActualizadoDto);
-        }else LOGGER.error("No fue posible actualizar los datos ya que el odontologo no se encuentra registrado");
-
-        return odontologoActualizadoDto;
+        Odontologo odontologoAActualizar = odontologoRepository.findById(odontologo.getId()).orElse(null);
+        OdontologoDto odontologoDtoActualizado = null;
+        if (odontologoAActualizar != null) {
+            odontologoAActualizar = odontologo;
+            odontologoRepository.save(odontologoAActualizar);
+            odontologoDtoActualizado = objectMapper.convertValue(odontologoAActualizar, OdontologoDto.class);
+            LOGGER.info("Dentist updated: {}", JsonPrinter.toString(odontologoDtoActualizado));
+        } else {
+            LOGGER.error("The dentist doesn´t exist");
+        }
+        return odontologoDtoActualizado;
     }
 
     @Override
-    public void eliminarOdontologo(Long id) {
-        odontologoRepository.deleteById(id);
-        LOGGER.warn("Se ha eliminado el odontologo con id {}", id);
+    public void eliminarOdontologo(Long id) throws ResourceNotFoundException {
+        if (buscarOdontologoPorId(id) != null) {
+            odontologoRepository.deleteById(id);
+            LOGGER.warn("Dentist with id {} deleted", id);
+        } else {
+            LOGGER.error("The dentist doesn´t exist");
+            throw new ResourceNotFoundException("The dentist doesn´t exist");
+        }
     }
-
 
 
 }
