+++ proyectoIntegrador/src/main/java/com/backend/proyectoIntegrador/service/impl/TurnoService.java
@@ -3,48 +3,67 @@ package com.backend.proyectoIntegrador.service.impl;
 import com.backend.proyectoIntegrador.dto.OdontologoDto;
 import com.backend.proyectoIntegrador.dto.PacienteDto;
 import com.backend.proyectoIntegrador.dto.TurnoDto;
-import com.backend.proyectoIntegrador.entity.Odontologo;
-import com.backend.proyectoIntegrador.entity.Paciente;
 import com.backend.proyectoIntegrador.entity.Turno;
+import com.backend.proyectoIntegrador.exceptions.BadRequestException;
+import com.backend.proyectoIntegrador.exceptions.ResourceNotFoundException;
 import com.backend.proyectoIntegrador.repository.TurnoRepository;
 import com.backend.proyectoIntegrador.service.ITurnoService;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.backend.proyectoIntegrador.utils.JsonPrinter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class TurnoService implements ITurnoService {
-    private static final Logger LOGGER = LoggerFactory.getLogger(PacienteService.class);
-    private TurnoRepository turnoRepository;
-    private ObjectMapper objectMapper;
-    @Override
-    public TurnoDto guardarTurno(Turno turno) {
-        Turno turnoNuevo = turnoRepository.save(turno);
-        PacienteDto pacienteDto = objectMapper.convertValue(turno.getPaciente(), PacienteDto.class);
-        OdontologoDto odontologoDto = objectMapper.convertValue(turno.getOdontologo(), OdontologoDto.class);
-        TurnoDto turnoNuevoDto = objectMapper.convertValue(turnoNuevo, TurnoDto.class);
-        turnoNuevoDto.setPacienteDto(pacienteDto);
-        turnoNuevoDto.setOdontologoDto(odontologoDto);
 
-        LOGGER.info("Date saved {}:", turno);
-        return turnoNuevoDto;
+    private static final Logger LOGGER = LoggerFactory.getLogger(TurnoService.class);
+    private final TurnoRepository turnoRepository;
+    private final PacienteService pacienteService;
+    private final OdontologoService odontologoService;
+
+    @Autowired
+    public TurnoService(TurnoRepository turnoRepository, PacienteService pacienteService, OdontologoService odontologoService) {
+        this.turnoRepository = turnoRepository;
+        this.pacienteService = pacienteService;
+        this.odontologoService = odontologoService;
+    }
+
+    @Override
+    public TurnoDto guardarTurno(Turno turno) throws BadRequestException {
+        TurnoDto turnoDto = null;
+        PacienteDto paciente = pacienteService.buscarPacientePorId(turno.getPaciente().getId());
+        OdontologoDto odontologo = odontologoService.buscarOdontologoPorId(turno.getOdontologo().getId());
+
+        if (paciente == null || odontologo == null) {
+            if (paciente == null && odontologo == null) {
+                LOGGER.error("The patient and dentist doesn´t exist in the data base");
+                throw new BadRequestException("The patient and dentist doesn´t exist in the data base");
+            } else if (paciente == null) {
+                LOGGER.error("The patient doesn´t exist in the data base");
+                throw new BadRequestException("The patient doesn´t exist in the data base");
+            } else {
+                LOGGER.error("The dentist doesn´t exist in the data base");
+                throw new BadRequestException("The dentist doesn´t exist in the data base");
+            }
+
+        } else {
+            turnoDto = TurnoDto.fromTurno(turnoRepository.save(turno));
+            LOGGER.info("New date saved: {}", JsonPrinter.toString(turnoDto));
+        }
+        return turnoDto;
     }
 
     @Override
     public List<TurnoDto> listarTodos() {
         List<Turno> turnos = turnoRepository.findAll();
         List<TurnoDto> turnoDtoList = turnos.stream()
-                .map(turno -> {
-                    Paciente paciente = turno.getPaciente();
-                    Odontologo odontologo = turno.getOdontologo();
-                    PacienteDto pacienteDto = objectMapper.convertValue(paciente, PacienteDto.class);
-                    OdontologoDto odontologoDto = objectMapper.convertValue(odontologo, OdontologoDto.class);
-                    return new TurnoDto(turno.getId(), turno.getFechaAsistencia().atStartOfDay(), pacienteDto, odontologoDto);
-                })
+                .map(TurnoDto::fromTurno)
                 .toList();
+
+        LOGGER.info("Dates list: {}", JsonPrinter.toString(turnoDtoList));
         return turnoDtoList;
     }
 
@@ -52,39 +71,40 @@ public class TurnoService implements ITurnoService {
     public TurnoDto buscarTurnoPorId(Long id) {
         Turno turnoBuscado = turnoRepository.findById(id).orElse(null);
         TurnoDto turnoDto = null;
+
         if (turnoBuscado != null) {
-            PacienteDto pacienteDto = objectMapper.convertValue(turnoBuscado.getPaciente(), PacienteDto.class);
-            OdontologoDto odontologoDto = objectMapper.convertValue(turnoBuscado.getOdontologo(), OdontologoDto.class);
-            turnoDto = objectMapper.convertValue(turnoBuscado, TurnoDto.class);
-            turnoDto.setOdontologoDto(odontologoDto);
-            turnoDto.setPacienteDto(pacienteDto);
-        } else LOGGER.info("The ID doesn´t exist");
+            turnoDto = TurnoDto.fromTurno(turnoBuscado);
+            LOGGER.info("Date founded: {}", JsonPrinter.toString(turnoDto));
+        } else LOGGER.info("The date doesn´t exist");
+
         return turnoDto;
     }
 
     @Override
     public TurnoDto actualizarTurno(Turno turno) {
         Turno turnoAActualizar = turnoRepository.findById(turno.getId()).orElse(null);
-        TurnoDto turnoActualizadoDto = null;
+        TurnoDto turnoDtoActualizado = null;
 
-        if(turnoAActualizar != null) {
+        if (turnoAActualizar != null) {
             turnoAActualizar = turno;
             turnoRepository.save(turnoAActualizar);
-
-            PacienteDto pacienteDto = objectMapper.convertValue(turnoAActualizar.getPaciente(), PacienteDto.class);
-            OdontologoDto odontologoDto = objectMapper.convertValue(turnoAActualizar.getOdontologo(), OdontologoDto.class);
-            turnoActualizadoDto = objectMapper.convertValue(turnoAActualizar, TurnoDto.class);
-            turnoActualizadoDto.setOdontologoDto(odontologoDto);
-            turnoActualizadoDto.setPacienteDto(pacienteDto);
-            LOGGER.info("Date updated {}:", turnoActualizadoDto);
+            turnoDtoActualizado = TurnoDto.fromTurno(turnoAActualizar);
+            LOGGER.warn("Date updated: {}", JsonPrinter.toString(turnoDtoActualizado));
         } else LOGGER.error("The date doesn´t exist");
 
-        return turnoActualizadoDto;
+        return turnoDtoActualizado;
     }
 
     @Override
-    public void eliminarTurno(Long id) {
-        turnoRepository.deleteById(id);
-        LOGGER.warn("The date with id {} was deleted", id);
+    public void eliminarTurno(Long id) throws ResourceNotFoundException {
+        if (buscarTurnoPorId(id) != null) {
+            turnoRepository.deleteById(id);
+            LOGGER.warn("Se ha eliminado el turno con id {}", id);
+        } else {
+            LOGGER.error("No se ha encontrado el turno con id {}", id);
+            throw new ResourceNotFoundException("No se ha encontrado el turno con id " + id);
+        }
     }
+
+
 }
